requirements.txt
Adafruit-Blinka==6.10.2
adafruit-circuitpython-amg88xx==1.2.8
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-mlx90640==1.2.3
adafruit-circuitpython-register==1.9.5
Adafruit-PlatformDetect==3.13.4
Adafruit-PureIO==1.1.9
appdirs==1.4.4
asn1crypto==0.24.0
astroid==2.1.0
asttokens==1.1.13
automationhat==0.2.0
beautifulsoup4==4.7.1
black==21.5b2
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
click==8.0.1
cmapy==0.6.6
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.0.2
fourletterphat==0.1.0
gpiozero==1.6.2
google-cloud-vision==3.1.2
h5py==2.8.0
html5lib==1.0.1
idna==2.6
importlib-metadata==4.5.0
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.3.1
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
matplotlib==3.4.2
mccabe==0.6.1
microdotphat==0.2.1
mlx90640-driver==1.1.2
mlx90640-driver-devicetree==1.1.0
mlx90640-driver-evb9064x==0.4.1
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.3
numpy==1.20.3
oauthlib==2.1.0
olefile==0.46
opencv-python==4.5.1.48
pantilthat==0.0.7
parso==0.3.1
pathspec==0.8.1
pexpect==4.6.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
piglow==1.2.5
pigpio==1.78
Pillow==8.2.0
pithermalcam==1.0.0
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pyftdi==0.53.1
pygame==1.9.4.post1
Pygments==2.3.1
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyserial==3.4
pysmbc==1.0.15.6
python-apt==1.8.4.3
python-dateutil==2.8.1
pytz==2021.1
pyusb==1.1.1
pyxdg==0.25
rainbowhat==0.1.0
regex==2021.4.4
reportlab==3.5.13
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scipy==1.6.3
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
sense-hat==2.2.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
ssh-import-id==5.7
sysv-ipc==1.1.0
thonny==3.3.6
toml==0.10.2
touchphat==0.0.1
twython==3.7.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
unicornhathd==0.0.4
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
zipp==3.4.1
